# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sample Pet Store App",
        "description": "This is a sample server for a pet store.",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:3000/api/v1",
            "description": "Localhost"
        }
    ],
...
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "components": {
        "schemas": {
            "id": {
                "components": {
                    "id": {
                        "type": "string"
                    }
                }
            },
...
            "users": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "userName": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "gmail": {
                        "type": "string"
                    },
                    "linkAvatar": {
                        "type": "string"
                    },
                    "statusActive": {
                        "type": "string"
                    },
                    "point": {
                        "type": "number"
                    },
                    "port": {
                        "type": "number"
                    },
                    "numberOfWins": {
                        "type": "number"
                    },
                    "numberOfLosses": {
                        "type": "number"
                    },
                    "resetPasswordToken": {
                        "type": "string"
                    },
                    "resetPasswordTokenExp": {
                        "type": "string"
                    }
                }
            },
...
        }
    },
    "paths": {
...
        "/users/{id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Lấy thông tin một user bất kì",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/users"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/delete/{id}": {
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "Xóa người dùng khỏi danh sách",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/users"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/edit/{id}": {
            "put": {
                "tags": [
                    "users"
                ],
                "summary": "Chỉnh sửa thông tin người dùng",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/id"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/editUser"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/users"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}